+++ android/src/main/kotlin/com/github/florent37/assets_audio_player/playerimplem/PlayerImplemExoPlayer.kt
@@ -6,6 +6,7 @@ import android.util.Log
 import com.github.florent37.assets_audio_player.AssetAudioPlayerThrowable
 import com.github.florent37.assets_audio_player.Player
 import com.google.android.exoplayer2.*
+import com.google.android.exoplayer2.C.AUDIO_SESSION_ID_UNSET
 import com.google.android.exoplayer2.Player.REPEAT_MODE_ALL
 import com.google.android.exoplayer2.Player.REPEAT_MODE_OFF
 import com.google.android.exoplayer2.extractor.DefaultExtractorsFactory
@@ -187,10 +188,20 @@ class PlayerImplemExoPlayer(
     }
 
     fun mapError(t: Throwable) : AssetAudioPlayerThrowable {
-        return if(t?.message?.contains("unable to connect",true) == true) {
-            AssetAudioPlayerThrowable.NetworkError(t)
-        } else {
-            AssetAudioPlayerThrowable.PlayerError(t)
+        return when {
+            t is ExoPlaybackException -> {
+                (t.cause as? HttpDataSource.InvalidResponseCodeException)?.takeIf { it.responseCode >= 400 }?.let {
+                    AssetAudioPlayerThrowable.UnreachableException(t)
+                } ?: let {
+                    AssetAudioPlayerThrowable.NetworkError(t)
+                }
+            }
+            t.message?.contains("unable to connect",true) == true -> {
+                AssetAudioPlayerThrowable.NetworkError(t)
+            }
+            else -> {
+                AssetAudioPlayerThrowable.PlayerError(t)
+            }
         }
     }
 
@@ -223,6 +234,7 @@ class PlayerImplemExoPlayer(
             this.mediaPlayer?.addListener(object : com.google.android.exoplayer2.Player.EventListener {
 
                 override fun onPlayerError(error: ExoPlaybackException) {
+                    Log.d("PLAYER","XXXX onPlayerError XXXX")
                     val errorMapped = mapError(error)
                     if (!onThisMediaReady) {
                         continuation.resumeWithException(errorMapped)
@@ -267,6 +279,8 @@ class PlayerImplemExoPlayer(
 
             mediaPlayer?.prepare(mediaSource)
         } catch (error: Throwable) {
+            Log.d("PLAYER","XXXX CATCH EXOPLAYER XXXX")
+
             if (!onThisMediaReady) {
                 continuation.resumeWithException(error)
             } else {
@@ -292,4 +306,7 @@ class PlayerImplemExoPlayer(
         mediaPlayer?.setPlaybackParameters(PlaybackParameters(playSpeed))
     }
 
+    override fun getSessionId(): Int? {
+        return mediaPlayer?.audioComponent?.audioSessionId?.takeIf { it != AUDIO_SESSION_ID_UNSET }
+    }
 }