+++ lib/src/playable.dart
@@ -1,11 +1,14 @@
 import 'package:assets_audio_player/assets_audio_player.dart';
 import 'package:flutter/foundation.dart';
 import 'package:uuid/uuid.dart';
+
 import 'utils.dart';
 
 class Playable {
   final Set<PlayerEditor> _currentlyOpenedIn = Set();
+
   Set<PlayerEditor> get currentlyOpenedIn => Set.from(_currentlyOpenedIn);
+
   void setCurrentlyOpenedIn(PlayerEditor player) {
     _currentlyOpenedIn.add(player);
   }
@@ -168,8 +171,10 @@ class Audio extends Playable {
   Metas _metas;
   Map<String, dynamic> _networkHeaders;
   final bool cached; //download audio then play it
+  final double playSpeed;
 
   Metas get metas => _metas;
+
   Map<String, dynamic> get networkHeaders => _networkHeaders;
 
   Audio._({
@@ -177,19 +182,27 @@ class Audio extends Playable {
     this.package,
     this.audioType,
     this.cached,
+    this.playSpeed,
     Map<String, dynamic> headers,
     Metas metas,
   })  : _metas = metas,
         _networkHeaders = headers;
 
-  Audio(this.path, {Metas metas, this.package})
-      : audioType = AudioType.asset,
+  Audio(
+    this.path, {
+    Metas metas,
+    this.package,
+    this.playSpeed,
+  })  : audioType = AudioType.asset,
         _networkHeaders = null,
         cached = false,
         _metas = metas;
 
-  Audio.file(this.path, {Metas metas})
-      : audioType = AudioType.file,
+  Audio.file(
+    this.path, {
+    Metas metas,
+    this.playSpeed,
+  })  : audioType = AudioType.file,
         package = null,
         _networkHeaders = null,
         cached = false,
@@ -200,6 +213,7 @@ class Audio extends Playable {
     Metas metas,
     Map<String, dynamic> headers,
     this.cached = false,
+    this.playSpeed,
   })  : audioType = AudioType.network,
         package = null,
         _networkHeaders = headers,
@@ -208,6 +222,7 @@ class Audio extends Playable {
   Audio.liveStream(
     this.path, {
     Metas metas,
+    this.playSpeed,
     Map<String, dynamic> headers,
   })  : audioType = AudioType.liveStream,
         package = null,
@@ -224,6 +239,7 @@ class Audio extends Playable {
           package == other.package &&
           audioType == other.audioType &&
           cached == other.cached &&
+          playSpeed == other.playSpeed &&
           metas == other.metas;
 
   @override
@@ -232,6 +248,7 @@ class Audio extends Playable {
       package.hashCode ^
       audioType.hashCode ^
       metas.hashCode ^
+      playSpeed.hashCode ^
       cached.hashCode;
 
   @override
@@ -263,6 +280,7 @@ class Audio extends Playable {
     String package,
     AudioType audioType,
     Metas metas,
+    double playSpeed,
     Map<String, dynamic> headers,
     bool cached,
   }) {
@@ -272,6 +290,7 @@ class Audio extends Playable {
       audioType: audioType ?? this.audioType,
       metas: metas ?? this._metas,
       headers: headers ?? this._networkHeaders,
+      playSpeed: playSpeed ?? this.playSpeed,
       cached: cached ?? this.cached,
     );
   }
@@ -283,7 +302,9 @@ class Playlist extends Playable {
   final List<Audio> audios = [];
 
   int _startIndex = 0;
+
   int get startIndex => _startIndex;
+
   set startIndex(int newValue) {
     if (newValue < this.audios.length) {
       _startIndex = newValue;