+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-## 2.0.5+6
+## 2.0.6+7
+
+* Cache now use `http` instead of `dio`
+* Added live tag on notification for LiveStream play (ios)
+* Added audio session id (android only)
+
+## 2.0.5+7
 
 * Added custom error handling (beta)
 * Dispose is now a future